Create prepare_vra_template_modified.ps1
Modified prepare vRA Template script to support Server 2016 Core. The standard script fails due to the extractZip function using a "shell.application" object to unpack the Java zip archive. In PowerShell 5 or later, a native cmdlet exists to unzip archives. Alternatively to this script you can follow the manual install instructions for the vRA agents which are available in the vRA 7.2 or earlier docs.